# src/main/resources/openapi/openapi-profile.yaml
openapi: 3.0.3
info:
  title: Microservices Social Network
  version: 1.0.11
  description: >-
    Here are written the API definition for a social network based on
    microservices technologies. This is a master thesis project.
  termsOfService: http://swagger.io/terms/
  contact:
    email: sergio.stefanizzi@finconsgroup.com
    name: Sergio Stefanizzi
servers:
  - url: http://localhost:8080/api/v1
    variables: {}
    description: Development server
...
paths:
...
  /admins/profiles:
    get:
      tags:
        - admins
      summary: List of all the profiles
      description: List of all the profiles
      operationId: findAllProfiles
      security:
        - bearerAuth: []
      parameters:
        - name: removedProfile
          in: query
          description: Show or not show removed profiles
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: List of all profiles found and returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admins/profiles/{profileId}:
    patch:
      tags:
        - admins
      summary: Block a profile
      description: Block a profile
      operationId: blockProfileById
      security:
        - bearerAuth: []
      parameters:
        - name: profileId
          in: path
          description: Id of the profile
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileAdminPatch'
      responses:
        '200':
          description: Profile has been successfully blocked!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admins/alerts:
    get:
      tags:
        - admins
      summary: List of all the alerts
      description: List of all the alerts
      operationId: findAllAlerts
      security:
        - bearerAuth: []
      parameters:
        - name: closedAlerts
          in: query
          description: Show or not show closed alerts
          required: true
          schema:
            type: boolean
        - name: adminId
          in: query
          description: Id of the admin profile in order to see its managed alerts
          schema:
            type: boolean
      responses:
        '200':
          description: List of all alerts found and returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/alerts/{alertId}:
    get:
      tags:
        - admins
      summary: Return an alert
      description: Return an alert
      operationId: findAlertById
      security:
        - bearerAuth: []
      parameters:
        - name: alertId
          in: path
          description: Id of the alert
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Alert found and returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - admins
      summary: Assign an alert to an admin
      description: Assign an alert to an admin
      operationId: updateAlertById
      security:
        - bearerAuth: []
      parameters:
        - name: alertId
          in: path
          description: Id of the alert
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertPatch'
      responses:
        '200':
          description: Alert has been successfully assigned!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
    BadRequest:
      description: Bad Request! Action not performed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorList'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Forbidden'
    Unauthenticated:
      description: Authentication is required!
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Unauthenticated'
...
    InternalServerError:
      description: Internal Error Server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        profile_name:
          type: string
          minLength: 8
          maxLength: 20
          pattern: '^(?=.{8,20}$)(?![_.])(?!.*[_.]{2})[a-zA-Z0-9._]+(?<![_.])$'
        bio:
          type: string
          maxLength: 150
          x-nonnull: '@com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)'
        picture_url:
          type: string
          minLength: 3
          maxLength: 2048
          x-nonnull: '@com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)'
          #x-url: '@org.hibernate.validator.constraints.URL'
          x-url: '@MyURL'
        is_private:
          type: boolean
        #TODO: con JWT non uso piu' account_id
        blocked_until:
          type: string
          format: date-time
          readOnly: true
          x-nonnull: '@com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)'
        account_id:
          type: integer
          format: int64
      required:
        - profile_name
        - is_private
        - account_id
...
    ProfileAdminPatch:
      type: object
      properties:
        blocked_until:
          type: string
          format: date-time
          x-future: '@Future'
          nullable: true
...
    Alert:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        created_by:
          type: integer
          format: int64
          nullable: true
        post_id:
          type: integer
          format: int64
          nullable: true
          x-nonnull: '@com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)'
        comment_id:
          type: integer
          format: int64
          nullable: true
          x-nonnull: '@com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)'
        reason:
          type: string
          minLength: 1
          maxLength: 2000
          example: reason example
        managed_by:
          type: integer
          format: int64
          nullable: true
          readOnly: true
          x-nonnull: '@com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)'
      required:
        - created_by
        - reason
    AlertPatch:
      type: object
      properties:
        managed_by:
          type: integer
          format: int64
      required:
        - managed_by
...
    ProfileList:
      type: array
      items:
        $ref: '#/components/schemas/Profile'
...
    AlertList:
      type: array
      items:
        $ref: '#/components/schemas/Alert'
    Error:
      type: object
      properties:
        error:
          type: string
          minLength: 1
          description: Error description
          example: Error description
    ErrorList:
      type: array
      items:
        $ref: '#/components/schemas/Error'
      description: List of errors
    NotFound:
      type: object
      properties:
        error:
          type: string
          minLength: 1
          example: Profile not found
      required:
        - error
    Unauthenticated:
      type: object
      properties:
        error:
          type: string
          minLength: 3
          example: Authentication is required!
    #Action is unauthenticated
    Forbidden:
      type: object
      properties:
        error:
          type: string
          minLength: 3
          example: Action is forbidden!
...
    InternalServerError:
      type: object
      properties:
        error:
          type: string
          minLength: 1
          example: Internal Server Error
      required:
        - error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT